Improves summarizer support for NPM license detection
NPM license may be declared as a string or array of strings
(even if the array contains only one license).
Additional support for legacy key 'licenses'
  https://docs.npmjs.com/files/package.json#license

Composer license may be declared as a string or array of strings
(even if the array contains only one license).
Also adds support for the NPM style legacy key 'licenses'
  https://getcomposer.org/doc/04-schema.md#license

Resolves #458

Signed-off-by: Tom Marble <tmarble@info9.net>