Changing github harvest recommendations to sort by best match.
The user/organization recommendations sorting for GitHub Harvest have
been changed from "repositories" to best match (the GitHub
default). This fixes an issue where an exact match is not the first
item in the results and a case where an exact match doesn't show up at
all in the recommendations (the user/org has few repos but their ID is
a substring of many other IDs; since we limit to first 100 results).

Fixes #661