get version from package.json instead of release
If the action is triggered by a release, `${{ github.event.release.tag_name }}` has the version.  If we manually run this action manually, `${{ github.event.release.tag_name }}` is empty.

To always have the version for tagging the created docker image and to pass to the webapp, this PR gets the version directly from the package.json file.

_NOTE: The package.json file must have the correct version when running manually._